Change file org to follow new pattern
