+++ src/components/ArticleCard/index.js
@@ -1,107 +1,1 @@
-// @flow
-import React, { PureComponent } from 'react'
-import Card from '../Card'
-import Text from '../Text'
-import Truncate from '../Truncate'
-
-import classNames from '../../utilities/classNames'
-import styled from '../styled'
-import css from './styles/ArticleCard.css.js'
-
-type Props = {
-  className?: string,
-  title?: string,
-  content?: string,
-  metaHeader?: any,
-  footer?: any,
-  titleLimit?: number,
-  contentLimit?: number,
-  titleSize?: number,
-  contentSize?: number,
-}
-
-/**
- * A enhanced wrapper that for Card, allowing for displaying a header, footer
- * and content inside a card
- */
-class ArticleCard extends PureComponent<Props> {
-  static defaultProps = {
-    titleLimit: 120,
-    contentLimit: 120,
-    titleSize: 13,
-    contentSize: 13,
-  }
-
-  renderTitle() {
-    const { title, titleLimit, titleSize } = this.props
-
-    return (
-      <div className="c-ArticleCard__title">
-        <Text size={titleSize}>
-          <Truncate limit={titleLimit} type="end">
-            {title}
-          </Truncate>
-        </Text>
-      </div>
-    )
-  }
-
-  renderContent() {
-    const { content, contentLimit, contentSize } = this.props
-
-    return (
-      <p className="c-ArticleCard__content">
-        <Text
-          size={contentSize}
-          className="c-ArticleCard__contentText"
-          block
-          muted
-        >
-          <Truncate limit={contentLimit} type="end">
-            {content}
-          </Truncate>
-        </Text>
-      </p>
-    )
-  }
-
-  renderFooter() {
-    const { footer } = this.props
-    return <div className="c-ArticleCard__footer">{footer}</div>
-  }
-
-  renderMetaHeader() {
-    const { metaHeader } = this.props
-    return <div className="c-ArticleCard__metaHeader">{metaHeader}</div>
-  }
-
-  render() {
-    const {
-      className,
-      title,
-      content,
-      metaHeader,
-      titleLimit,
-      contentLimit,
-      titleSize,
-      contentSize,
-      footer,
-      ...rest
-    } = this.props
-
-    const componentClassName = classNames('c-ArticleCard', className)
-
-    return (
-      <Card {...rest} className={componentClassName}>
-        {metaHeader && this.renderMetaHeader()}
-        {title && this.renderTitle()}
-        {content && this.renderContent()}
-        {footer && this.renderFooter()}
-      </Card>
-    )
-  }
-}
-
-const StyledCard = styled(ArticleCard)(css)
-
-export default StyledCard
+export default from './ArticleCard'
